refactor(tui::ui::components): move "GlobalListener" to own module
